pyproject.toml
[project]
name = "sakha"
version = "0.1.0"
description = "Sakha – Ward Assistant OpenEnv Environment"
requires-python = "==3.12.*"
dependencies = [
    "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "trl>=1.0.0",
    "accelerate>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
sakha-server = "server.app:main"
server = "server.app:main"
sakha = "sakha.__main__:main"

[build-system]
requires = ["uv_build>=0.8,<0.9"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--cov=src/sakha",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
filterwarnings = [
    "default::FutureWarning",
]
markers = [
    "slow: tests that are slow or integration-heavy",
    "integration: integration-level tests",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.15.8",
    "ty>=0.0.26",
]

[tool.ruff]
required-version = ">=0.15.8"
target-version = "py312"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "E402",   # allow deferred imports in inference script
]

[tool.ruff.lint.isort]
known-first-party = ["sakha"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
